Added user orders section
User can now view his orders, review system to be implemented

+++ database/migrations/2022_06_18_174338_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('title')->nullable();
-            $table->integer('price')->nullable();
+            $table->decimal('price', 7, 2)->nullable();
             $table->string('description')->nullable();
             $table->integer('quantity')->nullable();
             $table->string('image')->nullable();
