+++ database/migrations/2022_06_20_124810_create_carts_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('address')->nullable();
             $table->string('product_title')->nullable();
             $table->integer('quantity')->nullable();
-            $table->integer('price')->nullable();
+            $table->decimal('price', 7, 2)->nullable();
             $table->timestamps();
         });
     }
