+++ database/migrations/2022_06_21_133152_create_orders_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('address')->nullable();
             $table->string('product_name')->nullable();
             $table->integer('quantity')->nullable();
-            $table->integer('price')->nullable();
+            $table->decimal('price', 7, 2)->nullable();
             $table->string('status')->nullable();
             $table->timestamps();
         });
